+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+
+- For changes to the core logic and the API, see the [lowess](https://github.com/av746/lowess) and [fastLowess](https://github.com/av746/fastLowess) crates.
+
+### Added
+
+- Added support for new features in `fastLowess` v0.2.0.
+
+### Changed
+
+- Updated the documentation to reflect the latest changes.
+
 ## [0.1.1]
 
 ### Added